feat: adds lazy loading with intersection observer with images and fixes bug with files directory not being copied over

+++ packages/website/vite.config.js
@@ -39,10 +39,27 @@ function findAllBlogPostsWithMeta() {
 
 findAllBlogPostsWithMeta();
 
-fse.copySync(
-  path.resolve(__dirname, "..", "..", "posts", "files"),
-  path.resolve(__dirname, "public", "files")
-);
+// TODO: This should really be handled by the pages plugin
+function copyAllBlogPostsFilesToPublic() {
+  const postsPath = fastGlob.sync("../../posts/**/*.mdx");
+  for (const postPath of postsPath) {
+    const postDirectory = path.dirname(postPath);
+    const potentialFilesDirectory = path.resolve(postDirectory, "files");
+    if (fse.existsSync(potentialFilesDirectory)) {
+      fse.copySync(
+        potentialFilesDirectory,
+        path.resolve(__dirname, "public", "files")
+      );
+    }
+  }
+
+  fse.copySync(
+    path.resolve(__dirname, "..", "..", "posts", "files"),
+    path.resolve(__dirname, "public", "files")
+  );
+}
+
+copyAllBlogPostsFilesToPublic();
 
 export default defineConfig({
   plugins: [
